Fix likeness-seed race that flaked test_stack_query_respects_project_filter
Share the likeness pipeline quiesce/reseed helpers via tests/utils.py so the
seeded pairs can't collide with pipeline-computed ones.
