+++ tests/test_tag_health_api.py
@@ -16,9 +16,8 @@ from datetime import datetime, timedelta
 from fastapi.testclient import TestClient
 from PIL import Image
 from sqlalchemy import event as sa_event
-from sqlalchemy import func
 from sqlalchemy import insert as sa_insert
-from sqlmodel import delete, select
+from sqlmodel import select
 
 from pixlstash.db_models import (
     Picture,
@@ -28,13 +27,12 @@ from pixlstash.db_models import (
     PictureStack,
     Tag,
 )
-from pixlstash.db_models.picture_likeness import PictureLikenessQueue
 from pixlstash.db_models.tag_prediction import TagPrediction
 from pixlstash.db_models.tag_suggestion import TagSuggestion
 from pixlstash.db_models.tagger_run import TaggerRun
 from pixlstash.server import Server
 from pixlstash.utils.quality.anomaly_penalty import DEFAULT_TAG_PRECISION
-from tests.utils import upload_pictures_and_wait
+from tests.utils import seed_likeness_stable, upload_pictures_and_wait
 
 API = "/api/v1"
 
@@ -155,86 +153,6 @@ def _rebuild_and_wait(client, timeout_s=30):
     raise AssertionError("tag_health rebuild did not finish in time")
 
 
-def _wait_likeness_settled(server, timeout_s=180):
-    """Block until the background likeness pipeline has fully processed every
-    uploaded picture, so a manually-seeded ``PictureLikeness`` fixture will not
-    be clobbered underneath the board's ``mismatch`` read.
-
-    ``timeout_s`` is generous (CPU embeddings + likeness recompute for the
-    uploaded fixtures can take well over 30s on a loaded CI runner); the poll
-    returns as soon as the pipeline is quiescent, so a large cap costs nothing
-    on a fast machine and only avoids a spurious timeout under contention.
-
-    Uploading real pictures runs two background stages that mutate the
-    ``picturelikeness`` table: the likeness-parameter pass calls
-    ``LikenessParameterUtils.reset_likeness_for_pictures`` (which DELETEs every
-    pair touching a picture and re-queues it), then ``LikenessTask`` recomputes
-    real pairs. Both race with the tag-health reads and would delete/replace a
-    seeded near-duplicate pair, making ``mismatch`` non-deterministic (the value
-    even flips across test order because the recomputed real likeness straddles
-    ``MISMATCH_LIKENESS_THRESHOLD`` and depends on image size). Once every
-    picture is fully processed (embedding + likeness_parameters + perceptual_hash
-    all set), the queue is drained, and the pair table is stable across two
-    polls, the pipeline is quiescent: a subsequent atomic reseed (see
-    ``_seed_stable``) then stays put because the finder has no further work
-    (queue empty + pairs present).
-    """
-
-    def _snapshot(session):
-        total = int(session.exec(select(func.count()).select_from(Picture)).one())
-        ready = int(
-            session.exec(
-                select(func.count())
-                .select_from(Picture)
-                .where(
-                    Picture.image_embedding.is_not(None),
-                    Picture.likeness_parameters.is_not(None),
-                    Picture.perceptual_hash.is_not(None),
-                )
-            ).one()
-        )
-        queued = int(
-            session.exec(select(func.count()).select_from(PictureLikenessQueue)).one()
-        )
-        pairs = int(
-            session.exec(select(func.count()).select_from(PictureLikeness)).one()
-        )
-        return total, ready, queued, pairs
-
-    start = time.time()
-    prev_pairs = None
-    while time.time() - start < timeout_s:
-        total, ready, queued, pairs = server.vault.db.run_immediate_read_task(_snapshot)
-        # Quiescent only when every picture is fully processed and the queue is
-        # empty; stable across two polls confirms the last recompute wrote.
-        if total > 0 and ready == total and queued == 0 and pairs == prev_pairs:
-            return
-        prev_pairs = pairs
-        time.sleep(0.25)
-    raise AssertionError("likeness pipeline did not settle in time")
-
-
-def _seed_stable(server, seed_fn):
-    """Wait for the likeness pipeline to quiesce, then seed with a clean
-    ``picturelikeness`` slate in one atomic transaction.
-
-    Deletes any pipeline-computed pairs and drains the queue *inside* the seed
-    transaction so no observer ever sees an empty-table/empty-queue state (which
-    would re-trigger ``LikenessUtils.seed_queue``). After commit the queue is
-    empty and the seeded pairs are present, so ``MissingLikenessFinder`` has no
-    further work and the fixture stays authoritative. Returns whatever
-    ``seed_fn`` returns.
-    """
-    _wait_likeness_settled(server)
-
-    def _wrapped(session):
-        session.exec(delete(PictureLikeness))
-        session.exec(delete(PictureLikenessQueue))
-        return seed_fn(session)
-
-    return server.vault.db.run_task(_wrapped)
-
-
 def test_tag_health_aggregates_on_fixture_vault():
     temp_dir, client, server = _setup()
     try:
@@ -330,8 +248,8 @@ def test_tag_health_aggregates_on_fixture_vault():
 
         # Seed with a stable likeness slate: the background likeness pipeline
         # would otherwise delete/recompute the seeded near-duplicate pair and
-        # make `mismatch` non-deterministic (see _seed_stable).
-        _seed_stable(server, seed)
+        # make `mismatch` non-deterministic (see seed_likeness_stable).
+        seed_likeness_stable(server, seed)
 
         body = _rebuild_and_wait(client)
         assert body["computed_at"] is not None
@@ -391,7 +309,7 @@ def test_tag_health_same_stack_mismatch_and_no_double_count():
             )
             session.commit()
 
-        _seed_stable(server, seed)
+        seed_likeness_stable(server, seed)
 
         body = _rebuild_and_wait(client)
         rows = {r["tag"]: r for r in body["rows"]}
@@ -512,7 +430,7 @@ def test_tag_health_scoped_mismatch_excludes_out_of_scope_pairs():
             session.commit()
             return ps.id
 
-        set_id = _seed_stable(server, seed)
+        set_id = seed_likeness_stable(server, seed)
 
         # Vault-wide: both disagreeing pairs count.
         body = _rebuild_and_wait(client)
@@ -1145,7 +1063,7 @@ def test_tag_health_default_tag_merges_folds_across_all_signals():
             )
             session.commit()
 
-        _seed_stable(server, seed)
+        seed_likeness_stable(server, seed)
 
         body = _rebuild_and_wait(client)
         rows = {r["tag"]: r for r in body["rows"]}
